+++ linuxdoom-1.10/m_misc.h
@@ -31,11 +31,7 @@
 
 
 
-boolean
-M_WriteFile
-( char const*	name,
-  void*		source,
-  int		length );
+boolean M_WriteFile(char const* name, void* source, intptr_t length);
 
 int
 M_ReadFile
